+++ tests/integration/cases/empty_dict/Dart_declaration_style_var.dart
@@ -1,0 +1,1 @@
+var my_data = <String, dynamic>{};