+++ tests/integration/cases/empty_dict/FSharp_declaration_style_let_mutable.fs
@@ -1,0 +1,6 @@
+module Check
+
+type Val =
+    | FStr of string
+    | FMap of (string * Val) list
+let mutable my_data: Val = FMap []