+++ tests/integration/cases/int_key_dict/TypeScript_declaration_style_let.ts
@@ -1,0 +1,6 @@
+let my_data = {
+  "1": "one",
+  "2": "two",
+  "42": "answer",
+};
+export {};