+++ tests/integration/cases/int_list/Dart_declaration_style_const.dart
@@ -1,0 +1,5 @@
+const my_data = <int>[
+    1,
+    2,
+    3,
+];