+++ tests/integration/cases/int_list/Rust_sequence_array_decl_lazy_static.rs
@@ -1,0 +1,9 @@
+use std::sync::LazyLock;
+fn main() {
+    static my_data: LazyLock<[i32; 3]> = LazyLock::new(|| [
+        1,
+        2,
+        3,
+    ]);
+    let _ = my_data;
+}