+++ tests/integration/cases/int_list/Rust_sequence_tuple_decl_lazy_static.rs
@@ -1,0 +1,9 @@
+use std::sync::LazyLock;
+fn main() {
+    static my_data: LazyLock<(i32, i32, i32)> = LazyLock::new(|| (
+        1,
+        2,
+        3,
+    ));
+    let _ = my_data;
+}