+++ tests/integration/cases/int_list/Scala_declaration_style_var.scala
@@ -1,0 +1,7 @@
+object Check {
+var my_data = List[Int](
+    1,
+    2,
+    3,
+)
+}