+++ tests/integration/cases/scalar_int_large/Rust_declaration_style_lazy_static.rs
@@ -1,0 +1,5 @@
+use std::sync::LazyLock;
+fn main() {
+    static my_data: LazyLock<i64> = LazyLock::new(|| 2147483648i64);
+    let _ = my_data;
+}