+++ tests/integration/cases/scalar_null/Rust_declaration_style_lazy_static.rs
@@ -1,0 +1,5 @@
+use std::sync::LazyLock;
+fn main() {
+    static my_data: LazyLock<Option<()>> = LazyLock::new(|| None::<()>);
+    let _ = my_data;
+}